cli: Move rule formatting to PrettyFormatter #652
