+++ cli/zally/utils/result_printer_test.go
@@ -6,7 +6,6 @@ import (
 
 	"bytes"
 
-	"github.com/logrusorgru/aurora"
 	"github.com/zalando/zally/cli/zally/domain"
 	"github.com/zalando/zally/cli/zally/tests"
 )
@@ -20,59 +19,6 @@ func TestNewResultPrinter(t *testing.T) {
 
 		tests.AssertEquals(t, buffer, resultPrinter.buffer)
 		tests.AssertEquals(t, &formatter, resultPrinter.formatter)
-	})
-}
-
-func TestColorizeByTypeFunc(t *testing.T) {
-	var buffer bytes.Buffer
-	var formatter PrettyFormatter
-	resultPrinter := NewResultPrinter(&buffer, &formatter)
-
-	t.Run("Returns red when type is MUST", func(t *testing.T) {
-		result := resultPrinter.colorizeByTypeFunc("MUST")
-		tests.AssertEquals(t, aurora.Red("abcde"), result("abcde"))
-	})
-
-	t.Run("Returns brown when type is SHOULD", func(t *testing.T) {
-		result := resultPrinter.colorizeByTypeFunc("SHOULD")
-		tests.AssertEquals(t, aurora.Brown("abcde"), result("abcde"))
-	})
-
-	t.Run("Returns green when type is MAY", func(t *testing.T) {
-		result := resultPrinter.colorizeByTypeFunc("MAY")
-		tests.AssertEquals(t, aurora.Green("abcde"), result("abcde"))
-	})
-
-	t.Run("Returns cyan when type is HINT", func(t *testing.T) {
-		result := resultPrinter.colorizeByTypeFunc("HINT")
-		tests.AssertEquals(t, aurora.Cyan("abcde"), result("abcde"))
-	})
-
-	t.Run("Returns gray by default", func(t *testing.T) {
-		result := resultPrinter.colorizeByTypeFunc("WHATEVER")
-		tests.AssertEquals(t, aurora.Gray("abcde"), result("abcde"))
-	})
-}
-
-func TestPrintRule(t *testing.T) {
-	t.Run("Prints single rule", func(t *testing.T) {
-		var buffer bytes.Buffer
-		var formatter PrettyFormatter
-		resultPrinter := NewResultPrinter(&buffer, &formatter)
-
-		var rule domain.Rule
-		rule.Title = "Must Rule"
-		rule.Type = "MUST"
-		rule.Code = "166"
-		rule.IsActive = true
-		rule.URL = "https://example.com/rule"
-
-		resultPrinter.printRule(&rule)
-
-		tests.AssertEquals(
-			t,
-			"\x1b[31m166\x1b[0m \x1b[31mMUST\x1b[0m: Must Rule\n\thttps://example.com/rule\n\n",
-			buffer.String())
 	})
 }
